combine db.config.ts and general.config.ts to a single config.ts file for more readability
